Fixed Playerview

+++ app/src/main/java/com/example/cs2340a_team31/model/observers/PlayerObserver.java
@@ -1,5 +1,5 @@
 package com.example.cs2340a_team31.model.observers;
 
 public interface PlayerObserver {
-    void onPlayerPositionChanged(double x, double y);
+    void onPlayerPositionChanged(float x, float y);
 }
